changing navigation structure
