+++ lib/navigation/navigator.cpp
@@ -1,0 +1,36 @@
+#include "navigator.hpp"
+
+namespace hyped::navigation {
+
+Navigator::Navigator(){
+  //TODO: impement
+}
+
+void Navigator::navigate(){
+  /*
+  TODO: implement
+
+  main navigation control:
+
+  - look at new data
+  - preprocess
+  - calculate current trajecotory
+  - update current trajectory
+  
+  */
+}
+
+
+void Navigator::updateTrajectory(){
+
+  /*
+  TODO: implement
+
+  idk how ros works but make sure everyone who needs to knwo trajectory has it
+
+  */
+}
+
+
+
+}  // namespace hyped::navigation
