+++ lib/navigation/navigator.hpp
@@ -1,34 +1,19 @@
-#pragma once
+#include "consts.hpp"
 
-#include <core/types.hpp>
+namespace hyped::navigation{
 
-namespace hyped::navigation {
+class Navigator{
 
-struct Trajectory {
-  core::Float displacement;
-  core::Float velocity;
-  core::Float acceleration;
-};
+  public:
 
-inline bool operator==(const Trajectory &lhs, const Trajectory &rhs)
-{
-  return lhs.displacement == rhs.displacement && lhs.velocity == rhs.velocity
-         && lhs.acceleration == rhs.acceleration;
-}
+  Navigator();
 
-//Explicitly defined not equal to operator because gtest got sad
-inline bool operator!=(const Trajectory &lhs, const Trajectory &rhs)
-{
- return !(lhs == rhs);
-}
+  void navigate();
 
-inline Trajectory zero_trajectory = {0, 0, 0};
+  Trajectory trajectory_;
 
-template<typename SensorData>
-class INavigator {
- public:
-  virtual Trajectory currentTrajectory()             = 0;
-  virtual void update(const SensorData &sensor_data) = 0;
+  void updateTrajectory();
+
 };
 
 }  // namespace hyped::navigation
